Extract similar logic to AbstractHints
This also allows us to write an equals() method that is identical for
all Hints implementations (essentially just comparing the Hints maps).
